@@ -1,7 +1,7 @@
 package com.tutorial.crud.dto;
 
-import javax.validation.constraints.Min;
-import javax.validation.constraints.NotBlank;
+import jakarta.validation.constraints.Min;
+import jakarta.validation.constraints.NotBlank;
 
 public class ProductoDto {
 
