@@ -6,6 +6,9 @@ import com.nimbusds.jwt.JWTParser;
 import com.tutorial.crud.security.dto.JwtDto;
 import com.tutorial.crud.security.entity.UsuarioPrincipal;
 import io.jsonwebtoken.*;
+import io.jsonwebtoken.io.Decoders;
+import io.jsonwebtoken.security.Keys;
+import io.jsonwebtoken.security.SignatureException;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Value;
@@ -13,6 +16,7 @@ import org.springframework.security.core.Authentication;
 import org.springframework.security.core.GrantedAuthority;
 import org.springframework.stereotype.Component;
 
+import java.security.Key;
 import java.text.ParseException;
 import java.util.Date;
 import java.util.List;
@@ -36,17 +40,17 @@ public class JwtProvider {
                 .claim("roles", roles)
                 .setIssuedAt(new Date())
                 .setExpiration(new Date(new Date().getTime() + expiration * 180))
-                .signWith(SignatureAlgorithm.HS512, secret.getBytes())
+                .signWith(getSecret(secret))
                 .compact();
     }
 
     public String getNombreUsuarioFromToken(String token) {
-        return Jwts.parser().setSigningKey(secret.getBytes()).parseClaimsJws(token).getBody().getSubject();
+        return Jwts.parserBuilder().setSigningKey(getSecret(secret)).build().parseClaimsJws(token).getBody().getSubject();
     }
 
     public boolean validateToken(String token) {
         try {
-            Jwts.parser().setSigningKey(secret.getBytes()).parseClaimsJws(token);
+            Jwts.parserBuilder().setSigningKey(getSecret(secret)).build().parseClaimsJws(token);
             return true;
         } catch (MalformedJwtException e) {
             logger.error("token mal formado");
@@ -64,7 +68,7 @@ public class JwtProvider {
 
     public String refreshToken(JwtDto jwtDto) throws ParseException {
         try {
-            Jwts.parser().setSigningKey(secret.getBytes()).parseClaimsJws(jwtDto.getToken());
+            Jwts.parserBuilder().setSigningKey(getSecret(secret)).build().parseClaimsJws(jwtDto.getToken());
         } catch (ExpiredJwtException e) {
             JWT jwt = JWTParser.parse(jwtDto.getToken());
             JWTClaimsSet claims = jwt.getJWTClaimsSet();
@@ -76,10 +80,15 @@ public class JwtProvider {
                     .claim("roles", roles)
                     .setIssuedAt(new Date())
                     .setExpiration(new Date(new Date().getTime() + expiration))
-                    .signWith(SignatureAlgorithm.HS512, secret.getBytes())
+                    .signWith(getSecret(secret))
                     .compact();
         }
         return null;
+    }
+
+    private Key getSecret(String secret){
+        byte[] secretBytes = Decoders.BASE64URL.decode(secret);
+        return Keys.hmacShaKeyFor(secretBytes);
     }
 
 }
