@@ -2,6 +2,7 @@ package com.tutorial.crud.security.service;
 
 import com.tutorial.crud.security.entity.Usuario;
 import com.tutorial.crud.security.entity.UsuarioPrincipal;
+import com.tutorial.crud.security.repository.UsuarioRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.security.core.userdetails.UserDetailsService;
@@ -12,11 +13,11 @@ import org.springframework.stereotype.Service;
 public class UserDetailsServiceImpl implements UserDetailsService {
 
     @Autowired
-    UsuarioService usuarioService;
+    UsuarioRepository usuarioRepository;
 
     @Override
     public UserDetails loadUserByUsername(String nombreOrEmail) throws UsernameNotFoundException {
-        Usuario usuario = usuarioService.getByNombreUsuarioOrEmail(nombreOrEmail).get();
+        Usuario usuario = usuarioRepository.findByNombreUsuarioOrEmail(nombreOrEmail, nombreOrEmail).get();
         return UsuarioPrincipal.build(usuario);
     }
 }
